@@ -30,73 +30,68 @@ dependencies:
   flutter:
     sdk: flutter
 
-  biometric_storage: ^4.1.3
+  biometric_storage: ^5.0.1
 
-  flutter_slidable: ^0.6.0
-  provider: ^6.0.2
-  flutter_image_compress: ^1.1.0
-  path_provider: ^2.0.9
-  permission_handler: ^9.2.0
+  flutter_slidable: ^3.0.1
+  provider: ^6.1.2
+  flutter_image_compress: ^2.1.0
+  path_provider: ^2.1.2
+  permission_handler: ^11.3.0
   at_base2e15: ^1.0.0
-  youtube_player_flutter: ^8.0.0
-  receive_sharing_intent: ^1.4.5
+  youtube_player_flutter: ^9.0.0
+  receive_sharing_intent: 1.5.4
   http: ^0.13.4
   file_picker: ^4.5.0
-  easy_debounce: ^2.0.2+1
+  easy_debounce: ^2.0.3
   validators: ^3.0.0
-  latlong2: ^0.8.1
-  image_cropper: ^1.5.0
-  extended_image: ^8.0.2
-  image: ^3.1.1
-  showcaseview: ^1.1.1
+  latlong2: ^0.9.0
+  image_cropper: ^5.0.1
+  extended_image: ^8.2.0
+  image: ^4.1.7
+  showcaseview: ^2.0.3
   desktop_window: ^0.4.0
-  shared_preferences: ^2.0.7
+  shared_preferences: ^2.2.2
 
   html_editor_enhanced: ^2.5.1
-  flutter_widget_from_html: ^0.8.5
-  qr_flutter: ^4.0.0
-  share_plus: ^4.5.3
-  url_launcher: ^6.0.20
+  flutter_widget_from_html: ^0.14.11
+  qr_flutter: ^4.1.0
+  share_plus: ^7.2.2
+  url_launcher: ^6.2.5
 
-  at_common_flutter: ^2.0.11
-  at_onboarding_flutter: ^6.0.3
-  at_client_mobile: ^3.2.9
-  at_contact: ^3.0.7
-  at_lookup: ^3.0.35
-  at_contacts_flutter: ^4.0.9
-  at_commons: ^3.0.43
-  at_follows_flutter: ^3.0.9
-  at_location_flutter: ^3.1.8
-  at_sync_ui_flutter: ^1.0.8
+  at_common_flutter: ^2.0.13
+  at_onboarding_flutter: ^6.1.7
+  at_client_mobile: ^3.2.15
+  at_contact: ^3.0.8
+  at_lookup: ^3.0.46
+  at_contacts_flutter: ^4.0.14
+  at_commons: ^4.0.3
+  at_follows_flutter: ^3.0.12
+  at_location_flutter: ^3.1.12
+  at_sync_ui_flutter: ^1.0.11
   uni_links: ^0.5.1
   internet_connection_checker: ^1.0.0+1
-  at_server_status: ^1.0.3
-  flutter_dotenv: ^5.0.2
+  at_server_status: ^1.0.4
+  flutter_dotenv: ^5.1.0
 
-  new_version: ^0.3.0
-  another_flushbar: ^1.10.28
+  new_version: ^0.3.1
+  another_flushbar: ^1.12.30
 
 dependency_overrides:
-  fwfh_text_style: ^2.22.8+1
-  flutter_local_notifications: 9.7.1
-  permission_handler: ^10.2.0
-  flutter_inappwebview: 5.4.4+3
-  file_picker: ^5.2.5
-  webview_flutter: ^4.0.1
-  fwfh_webview: 0.7.0
-  geolocator: ^9.0.2
-  path: ^1.8.3
-  numberpicker: ^2.1.2
-  # at_onboarding_flutter:
-  #   git:
-  #     url: https://github.com/atsign-foundation/at_widgets.git
-  #     path: packages/at_onboarding_flutter
-  #     ref: trunk
+  http: ^1.2.1
+  package_info_plus: ^5.0.0
+  web: ^0.5.0
+  file_picker: ^6.1.1
+  flutter_inappwebview: ^6.0.0
+#  at_onboarding_flutter:
+#    git:
+#     url: https://github.com/atsign-foundation/at_widgets.git
+#     path: packages/at_onboarding_flutter
+#     ref: trunk
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  msix: ^3.3.1
+  msix: ^3.16.7
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
